refactor: load stores in Home page component
