+++ mobile/src/app/(tabs)/Home/index.tsx
@@ -14,7 +14,6 @@ import Icon from 'react-native-vector-icons/Feather';
 import CarouselSection from '../../../components/CarouselSection';
 import CircularCarousel from '../../../components/CircularCarousel';
 import { CarouselItem } from '../../../components/CarouselSection/types';
-import { fetchLojas } from '@/app/lojas';
 import { styles } from './styles';
 
 interface Product {
@@ -37,13 +36,6 @@ const promotions: CarouselItem[] = Array.from({ length: 5 }).map((_, i) => ({
   image: `https://picsum.photos/seed/promo${i}/300/200`,
 }));
 
-const fallbackNearbyPromotions: CarouselItem[] = Array.from({ length: 6 }).map(
-  (_, i) => ({
-    id: `np${i}`,
-    title: `Oferta ${i + 1}`,
-    image: `https://picsum.photos/seed/nearby${i}/300/300`,
-  }),
-);
 
 function generateProducts(count: number): Product[] {
   return Array.from({ length: count }).map((_, i) => {
@@ -59,12 +51,14 @@ function generateProducts(count: number): Product[] {
   });
 }
 
-export default function HomeScreen() {
+export interface HomeScreenProps {
+  nearbyPromotions: CarouselItem[];
+}
+
+export default function HomeScreen({ nearbyPromotions }: HomeScreenProps) {
   const [location, setLocation] = useState('Obtendo localização...');
   const [products, setProducts] = useState<Product[]>([]);
   const [loading, setLoading] = useState(false);
-  const [nearbyPromotions, setNearbyPromotions] =
-    useState<CarouselItem[]>(fallbackNearbyPromotions);
 
   const loadMore = useCallback(() => {
     if (loading) return;
@@ -95,19 +89,6 @@ export default function HomeScreen() {
       }
     })();
     loadMore();
-    fetchLojas()
-      .then(lojas => {
-        setNearbyPromotions(
-          lojas.map((loja, i) => ({
-            id: String(loja.id),
-            title: loja.nome,
-            image: loja.imagem ?? `https://picsum.photos/seed/nearby${i}/300/300`,
-          })),
-        );
-      })
-      .catch(() => {
-        setNearbyPromotions(fallbackNearbyPromotions);
-      });
   }, [loadMore]);
 
   const renderProduct = ({ item }: ListRenderItemInfo<Product>) => (
